updates build-first-slack-bot-python tutorial for modern Slack Apps
the content and screenshots are updated to push away from Slack custom integrations
and towards Slack Apps. Custom integrations are less powerful and represent a
dead end in terms of further development in the Slack Platform.

one significant code change is the removal of the separate script used to find
the bot user ID, as there is an API method that can be used to get it at
runtime.