Improve url detection
The previous implementation used Linux regular expressions. This was
sufficient as a MVP but it was not accurate enough for corner cases.
After doing some more research, it seemed as if using a HTML parsing
library would be more efficient for this purpose. As such, the shell
command has been scrapped away in favor of a more elaborate approach
for detecting urls.

- Implement skeleton for extract_urls
- Detect html and markdown files
- Use bs4 for parsing html
- Convert markdown for bs4 parsing
- Remove use of urlin.txt and urlout.txt
- Remove unnecessary global vars