Update 180309-flask-docker-macos.markdown
Hey,

I know this is a slightly older document, however I found it really useful in getting started with Docker, thank you! 

I've updated where requirements.txt is created as I was unable to create the image until this was done. As I created an empty requirements.txt to start with I found it failed because then Flask was unavailable and I needed to rebuild the image. :) 

Donna